+++ librz/il/il_events.c
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2021 deroad <wargio@libero.it>
 // SPDX-License-Identifier: LGPL-3.0-only
 
-#include <rz_il/rzil_vm_events.h>
+#include <rz_il/rz_il_events.h>
 
 /**
  * Frees an RzILEvent struct
