+++ librz/il/theory_effect.c
@@ -2,9 +2,8 @@
 // SPDX-FileCopyrightText: 2021 heersin <teablearcher@gmail.com>
 // SPDX-License-Identifier: LGPL-3.0-only
 
-#include <rz_il/rzil_opcodes.h>
-#include <rz_il/vm_layer.h>
-#include <rz_il/rzil_vm.h>
+#include <rz_il/rz_il_opcodes.h>
+#include <rz_il/rz_il_vm.h>
 
 static RzILEvent *il_event_new_write_from_var(RzILVM *vm, RzILVar *var, RzILVal *new_val) {
 	rz_return_val_if_fail(vm && var && new_val, NULL);
