+++ librz/il/theory_mem.c
@@ -2,9 +2,8 @@
 // SPDX-FileCopyrightText: 2021 heersin <teablearcher@gmail.com>
 // SPDX-License-Identifier: LGPL-3.0-only
 
-#include <rz_il/rzil_vm.h>
-#include <rz_il/rzil_opcodes.h>
-#include <rz_il/vm_layer.h>
+#include <rz_il/rz_il_vm.h>
+#include <rz_il/rz_il_opcodes.h>
 
 void *rz_il_handler_load(RzILVM *vm, RzILOpBitVector *op, RzILPureType *type) {
 	rz_return_val_if_fail(vm && op && type, NULL);
