+++ librz/include/rz_il.h
@@ -1,8 +1,7 @@
 #ifndef RZ_IL_H
 #define RZ_IL_H
 
-#include <rz_il/vm_layer.h>
-#include <rz_il/rzil_vm.h>
-#include <rz_il/rzil_opcodes.h>
+#include <rz_il/rz_il_vm.h>
+#include <rz_il/rz_il_opcodes.h>
 
 #endif // RZ_IL_H
